Started Implementation on Main Encoding Function and Metadata
Started Implementation of Main Encoding Function and Metadata Creation Function.
The Metadata is fixed in size and encodes a magic byte, the number of divisions and will encode other information in the future.
The Encoding functions creates frames until all data is exausthed, then pads the final frame with 0s until it is filled. For now, frame are saved as PNGs on the disk.