Moved to FFMPEG Lib and a more fault tolerant encoding
Used FFMPEG Library (through skvideo) directly, as it is faster than using the OpenCV Wrapper and offers more control on the output. Data is now encoded using only the most significant half of every byte (the first 4 bits) and padding with 0s. This doubles the file size but makes decoding possibile, as all bytes now have a CRC code intrinsicly added to them.