Encoding function implemented
Encoding function has been implemented. It is more then x10 faster then the original implementation of the same function. It now uses numpy function to resize a smaller matrix, with only 1 pixel for each division, to the full video size. This reduces the number of for cycles from 4 to 1 in the encode functions.