Added some more checks to the user input
Added a check for the output path of the encode operation. If path does not end with .mp4, which is necessary for FFMPEG to understand the format of the video it must create, a .mp4 extension will be added to the path, even if the may make the program crash if the path is to a directory and not to a file.