Added Automatic Extension detector & Changed Heuristic Strategy
Added a check at the end of the decode function that tries to guess the extension of the file analyzing magic numbers and headers. Also changed heuristic strategy: it now cycles pixels from top to bottom, instead of from left to right. This is because many filetype have zero-bytes as the first row, so an heurisitc check may fail. Top to bottom works better (tried on STLs and PNGs)